requirements.txt
streamlit
youtube-transcript-api
openai-whisper
python-docx
yt-dlp
torch